requirements.txt
# ── Core Framework ──────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20

# ── LangChain + LLM Providers ──────────────────────
# All 1.x family — must stay consistent
langchain>=1.0.0,<2.0.0
langchain-openai>=1.0.0,<2.0.0
langchain-community>=0.4.0,<1.0.0
langchain-core>=1.2.18,<2.0.0
langchain-groq>=0.2.0

# ── Vector Store (Local) ───────────────────────────
faiss-cpu==1.9.0.post1
scikit-learn==1.5.2

# ── Azure Services (optional — for production) ────
azure-search-documents==11.6.0b8
azure-identity==1.19.0
redis==5.2.1

# ── Testing ────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
